refactor: Cleanup imports for simple tree-shaking
