+++ app/scripts/PixiTrack.js
@@ -3,10 +3,10 @@ import slugid from 'slugid';
 
 import Track from './Track';
 
-import { colorToHex } from './utils';
+import colorToHex from './utils/color-to-hex';
 
 // Configs
-import { GLOBALS } from './configs';
+import GLOBALS from './configs/globals';
 
 /**
  * Format a resolution relative to the highest possible resolution.
