+++ app/scripts/Track.js
@@ -2,7 +2,7 @@ import { scaleLinear } from 'd3-scale';
 import { fake as fakePubSub } from './hocs/with-pub-sub';
 
 // Services
-import { isWithin } from './utils';
+import isWithin from './utils/is-within';
 
 class Track {
   constructor(context) {
