+++ app/scripts/gosling-exports.js
@@ -9,22 +9,21 @@ export { default as TiledPixiTrack } from './TiledPixiTrack';
 export { default as SVGTrack } from './SVGTrack';
 export { default as ViewportTrackerHorizontal } from './ViewportTrackerHorizontal';
 export { default as HeatmapTiledPixiTrack } from './HeatmapTiledPixiTrack';
+
 /**
  * Data Fetcher
  */
-export { DataFetcher } from './data-fetchers';
+export { default as DataFetcher } from './data-fetchers/DataFetcher';
 /**
  * Utils
  */
 export { fake as fakePubSub } from './hocs/with-pub-sub';
-export { tileProxy } from './services';
-export { DenseDataExtrema1D } from './utils';
-export {
-  absToChr,
-  chrToAbs,
-  colorToHex,
-  chromInfoBisector,
-  pixiTextToSvg,
-  svgLine,
-  showMousePosition,
-} from './utils';
+export { default as tileProxy } from './services/tile-proxy';
+export { default as DenseDataExtrema1D } from './utils/DenseDataExtrema1D';
+export { default as absToChr } from './utils/abs-to-chr';
+export { default as chrToAbs } from './utils/chr-to-abs';
+export { default as colorToHex } from './utils/color-to-hex';
+export { default as chromInfoBisector } from './utils/chrom-info-bisector';
+export { default as pixiTextToSvg } from './utils/pixi-text-to-svg';
+export { default as svgLine } from './utils/svg-line';
+export { default as showMousePosition } from './utils/show-mouse-position';
